Added widget switcher in the main UI to show either the Conversation or the Files Changed, like in GH's UI Added a way to add comments to a PR directly
Fixes #9
