+++ Source/GitHubTools/Private/Widgets/SGitHubToolsPRReviewThreadTableRow.h
@@ -20,7 +20,6 @@ public:
 
 private:
     FSlateColor GetBorderBackgroundColor() const;
-    TSharedRef< ITableRow > GenerateCommentRow( FGithubToolsPullRequestCommentPtr item, const TSharedRef< STableViewBase > & owner_table );
     FReply OnResolveConversationClicked();
     FText GetCollapsedButtonText() const;
     FReply OnCollapsedButtonClicked();
